Move supporting JSON viewer code
