+++ frontend/src/__generated/ve/components/JsonViewer/JsonViewerObject.css.js
@@ -1,0 +1,13 @@
+// src/components/JsonViewer/JsonViewerObject.css.ts
+var attributeAction = "_19or1k13";
+var attributeActions = "_19or1k12";
+var attributeLine = "_19or1k11";
+var buttonLink = "_19or1k14";
+var line = "_19or1k10";
+export {
+  attributeAction,
+  attributeActions,
+  attributeLine,
+  buttonLink,
+  line
+};
