+++ frontend/src/pages/LogsPage/utils.ts
@@ -2,10 +2,6 @@ import moment from 'moment'
 import { stringify } from 'query-string'
 import { DateTimeParam, encodeQueryParams, StringParam } from 'use-query-params'
 
-import {
-	AndOrExpression,
-	SearchExpression,
-} from '@/components/Search/Parser/listener'
 import { TIME_FORMAT } from '@/components/Search/SearchForm/constants'
 import { DEFAULT_OPERATOR } from '@/components/Search/SearchForm/utils'
 import {
@@ -24,81 +20,6 @@ export const isSignificantDateRange = (startDate: Date, endDate: Date) => {
 		moment(startDate).format(TIME_FORMAT) !==
 		moment(endDate).format(TIME_FORMAT)
 	)
-}
-
-const bodyKey = ReservedLogKey['Message']
-
-const EXISTS_PLACEHOLDER_VALUE = 'EXISTS'
-
-export const findMatchingLogAttributes = (
-	queryParts: Array<SearchExpression | AndOrExpression>,
-	logAttributes: object | string,
-	matchingAttributes: any = {},
-	attributeKeyBase: string[] = [],
-): { [key: string]: { match: string; value: string } } => {
-	if (!queryParts?.length || !logAttributes) {
-		return {}
-	}
-
-	const bodyQueryValues = queryParts
-		.filter(
-			(term): term is SearchExpression =>
-				'key' in term && term.key === bodyKey,
-		)
-		.map((term) => term.value.replace(/^"|"$/g, ''))
-
-	Object.entries(logAttributes).forEach(([key, value]) => {
-		const isString = typeof value === 'string'
-
-		if (!isString) {
-			findMatchingLogAttributes(queryParts, value, matchingAttributes, [
-				...attributeKeyBase,
-				key,
-			])
-			return
-		}
-
-		const bodyQueryValue = bodyQueryValues.find(
-			(bodyQueryValue) =>
-				value.toLowerCase().indexOf(bodyQueryValue.toLowerCase()) > -1,
-		)
-
-		let matchingAttribute: string | undefined = undefined
-		if (bodyQueryValue && key === bodyKey) {
-			matchingAttribute = bodyQueryValue
-		} else {
-			const fullKey = [...attributeKeyBase, key].join('.')
-
-			queryParts.some((term) => {
-				if (!('key' in term)) {
-					return false
-				}
-
-				const queryKey = term.key.toLowerCase()
-				const queryOpertor = term.operator
-				const queryValue = term.value?.toLowerCase()
-
-				if (queryKey === fullKey) {
-					// TODO: figure out why operator is 'NOTEXISTS' without spaces
-					// so we can use the constants
-					matchingAttribute = ['EXISTS', 'NOTEXISTS'].includes(
-						queryOpertor.toUpperCase(),
-					)
-						? EXISTS_PLACEHOLDER_VALUE
-						: queryValue
-				}
-			})
-		}
-
-		if (!!matchingAttribute) {
-			matchingAttributes[[...attributeKeyBase, key].join('.')] = {
-				match: matchingAttribute,
-				value,
-			}
-		}
-	})
-
-	return matchingAttributes
 }
 
 export const buildSessionParams = ({
